DSEGOG-106 fixing column state reset bug
Caused by availableColumns refreshing when the user modified the table in some way
