+++ src/table/table.component.test.tsx
@@ -71,6 +71,7 @@ describe('Table', () => {
       onDragEnd,
       onColumnWordWrapToggle,
       openFilters,
+      filteredChannelNames: [],
     };
   });
