+++ src/table/table.component.tsx
@@ -61,6 +61,7 @@ export interface TableProps {
   onDragEnd: (result: DropResult) => void;
   onColumnClose: (column: string) => void;
   openFilters: (headerName: string) => void;
+  filteredChannelNames: string[];
 }
 
 const Table = React.memo((props: TableProps): React.ReactElement => {
@@ -82,6 +83,7 @@ const Table = React.memo((props: TableProps): React.ReactElement => {
     onDragEnd,
     onColumnClose,
     openFilters,
+    filteredChannelNames,
   } = props;
 
   /*
@@ -202,7 +204,7 @@ const Table = React.memo((props: TableProps): React.ReactElement => {
                                       ...stickyColumnStyles,
                                     }
                                   : columnStyles;
-                                const { channelInfo, filtered } =
+                                const { channelInfo } =
                                   column as ColumnInstance;
 
                                 return (
@@ -225,7 +227,9 @@ const Table = React.memo((props: TableProps): React.ReactElement => {
                                       columnStates[dataKey]?.wordWrap ?? false
                                     }
                                     onToggleWordWrap={onColumnWordWrapToggle}
-                                    isFiltered={filtered}
+                                    isFiltered={filteredChannelNames.includes(
+                                      dataKey
+                                    )}
                                     openFilters={openFilters}
                                   />
                                 );
